datafeeder: refactor for higher cohesion of services
