+++ datafeeder/src/main/java/org/georchestra/datafeeder/batch/analysis/DataUploadStateInitializer.java
@@ -21,6 +21,7 @@ package org.georchestra.datafeeder.batch.analysis;
 import java.util.Objects;
 import java.util.UUID;
 
+import org.georchestra.datafeeder.batch.service.DataUploadAnalysisService;
 import org.georchestra.datafeeder.model.DataUploadJob;
 import org.georchestra.datafeeder.model.DatasetUploadState;
 import org.georchestra.datafeeder.model.JobStatus;
